Add price column to CarsAndOrders model
